Renaming AuthControllerTest -> AuthServiceTest
